feat: handle not enough payment

+++ src/main/java/com/tiemcheit/tiemcheitbe/repository/PaymentRepo.java
@@ -10,7 +10,7 @@ public interface PaymentRepo extends JpaRepository<Payment, Long> {
     Optional<Payment> findTop1ByUsernameOrderByOrderDateDesc(String username);
 
     @Query(value = "SELECT * FROM payments p WHERE p.username = :username AND p.total_price = :amount ORDER BY p.order_date DESC LIMIT 1", nativeQuery = true)
-    Optional<Payment> findMatchingPayment(String username, Long amount);
+    Optional<Payment> findMatchingPayment(String username, Double amount);
 
     void deleteByUsername(String username);
 }
